% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/latentclassification.R
\name{latentclassfication}
\alias{latentclassfication}
\title{Title latentclassification}
\usage{
latentclassfication(z, w, n_cluster)
}
\arguments{
\item{z}{z estimate from lsirm12pl function}

\item{w}{w estimate from lsirm12pl function}

\item{n_cluster}{number of cluster}
}
\value{
set of plots
}
\description{
Title latentclassification
}
