% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/latentclassification-package.R
\docType{package}
\name{latentclassification-package}
\alias{latentclassification}
\alias{latentclassification-package}
\title{latentclassification: This package helps drawing plots of classification in latent space}
\description{
When analyzing the result of latent space item response model, it is important to look at the components of latent space. In latent space, there are respondents and items. One of the goals to fit a latent space item respons model is to see the interactions and relationships between respondents group and item group. By setting the dimension of unobserved space 2, we can visualize and observe the interaction between respondents and items. In this package, we provide an easy tool that can visualize latent space. First, we starts with the kmeans clustering and provide silhouette plot so we can determine the number of group. Also, this plot package provides the results of other classification methods such as hierarchical clustering and PCA.
}
\keyword{internal}
