% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/latentclassification.R
\name{latentclassification}
\alias{latentclassification}
\title{Latentclassification}
\usage{
latentclassification(z, w, n_cluster)
}
\arguments{
\item{z}{z estimate from lsirm12pl function. It's nx2 matrix}

\item{w}{w estimate from lsirm12pl function. It's px2 matrix}

\item{n_cluster}{number of cluster. It's constant}
}
\value{
set of plots
}
\description{
Help doing classification in the latent space and drawing a plot.
}
